refactor(mlx90632d): support extended range at init
