+++ src/components/i2c/WipperSnapper_I2C.h
@@ -50,7 +50,7 @@
 #include "drivers/WipperSnapper_I2C_Driver_MAX17048.h"
 #include "drivers/WipperSnapper_I2C_Driver_MCP3421.h"
 #include "drivers/WipperSnapper_I2C_Driver_MCP9808.h"
-#include "drivers/WipperSnapper_I2C_Driver_MLX90632.h"
+#include "drivers/WipperSnapper_I2C_Driver_MLX90632D.h"
 #include "drivers/WipperSnapper_I2C_Driver_MPL115A2.h"
 #include "drivers/WipperSnapper_I2C_Driver_MPRLS.h"
 #include "drivers/WipperSnapper_I2C_Driver_MS8607.h"
@@ -181,7 +181,8 @@ private:
   WipperSnapper_I2C_Driver_LTR390 *_ltr390 = nullptr;
   WipperSnapper_I2C_Driver_MCP3421 *_mcp3421 = nullptr;
   WipperSnapper_I2C_Driver_MCP9808 *_mcp9808 = nullptr;
-  WipperSnapper_I2C_Driver_MLX90632D *_mlx90632 = nullptr;
+  WipperSnapper_I2C_Driver_MLX90632D *_mlx90632d = nullptr;
+  WipperSnapper_I2C_Driver_MLX90632D *_mlx90632d_ext = nullptr;
   WipperSnapper_I2C_Driver_MPL115A2 *_mpl115a2 = nullptr;
   WipperSnapper_I2C_Driver_MPRLS *_mprls = nullptr;
   WipperSnapper_I2C_Driver_MS8607 *_ms8607 = nullptr;
